Move helpers located in `_internal` folder to main `helpers` folder
