+++ lib/helpers/remove-configuration-html-comments.js
@@ -4,7 +4,7 @@
 //   <!-- template-lint no-triple-curlies=false -->
 //   <!-- template-lint enabled=false -->
 
-const AstNodeInfo = require('../../helpers/ast-node-info');
+const AstNodeInfo = require('./ast-node-info');
 
 module.exports = function () {
   return class RemoveConfigurationHtmlCommentsRule {
