+++ lib/helpers/scope.js
