+++ lib/rules/_base.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const DEPRECATED_RULES = require('../helpers/deprecated-rules');
-const Scope = require('./_internal/scope');
+const Scope = require('../helpers/scope');
 
 // Constant to return from AST handlers so we know to remove/unregister them
 const REMOVE_HANDLER = {};
